requirements/base.txt
Django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-filter==21.1
psycopg2==2.9.1
channels==3.0.4
channels-redis==3.3.1
Pillow==8.4.0
uWSGI>=2.0.19.1,<2.1
celery==5.1.2
redis==3.5.3
djangorestframework-simplejwt==5.0.0